Logout adjustment and operation summary.

+++ app/Http/Controllers/Auth/AuthController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\LoginResquest;
 use App\Models\User;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
+use Illuminate\Support\Str;
 use Laravel\Sanctum\PersonalAccessToken;
 
 class AuthController extends Controller
@@ -36,13 +37,21 @@ class AuthController extends Controller
     {
         $requestToken = $request->header('authorization');
 
+        $array = Str::of($requestToken)->explode('|');
+
         $token = (new PersonalAccessToken())
-            ->findToken(substr($requestToken, 10, strlen($requestToken)));
+            ->findToken($array->offsetGet(1));
+
+        if (!$token){
+            return response()->json([
+                'message' => 'Token has already been revoked.'
+            ], 500);
+        }
 
         $token->delete();
 
         return response()->json([
-            'message' => 'Roken Revoked.'
+            'message' => 'Token Revoked.'
         ], 200);
     }
 }
